wired log in, register, and file upload to modals
