+++ src/main/java/com/artiscene/controllers/HomeController.java
@@ -1,9 +1,17 @@
 package com.artiscene.controllers;
 
+import com.artiscene.models.User;
+import com.artiscene.repositories.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+
+import javax.validation.Valid;
 
 /**
  * Created by HKoehler on 2/17/17.
@@ -11,9 +19,43 @@ import org.springframework.web.bind.annotation.PostMapping;
 
 @Controller
 public class HomeController {
+    private UserRepository repository;
+    private PasswordEncoder encoder;
+
+    @Autowired
+    public HomeController(UserRepository repository, PasswordEncoder encoder){
+        this.repository=repository;
+        this.encoder=encoder;
+    }
 
     @GetMapping("/")
     public String homepage(Model model) {
+        model.addAttribute("user", new User());
         return "home";
+    }
+
+//    @GetMapping("/register")
+//    public String registerPage(Model model) {
+//        model.addAttribute("user", new User());
+//        return "forms/register";
+//    }
+
+    @PostMapping("/register")
+    public String registerUser(@Valid User user, Errors validation, Model model, @RequestParam(name="password_confirm") String passwordConfirmation) {
+
+        if (!passwordConfirmation.equals(user.getPassword())) {
+            validation.rejectValue("password", "user.password", "Your passwords do not match");
+        }
+
+        if (validation.hasErrors()) {
+            model.addAttribute("errors", validation);
+            model.addAttribute("user", user);
+            return "redirect:/";
+        }
+        String hashedPassword = encoder.encode(user.getPassword());
+        user.setPassword(hashedPassword);
+        repository.save(user);
+        return "redirect:/";
+
     }
 }
