+++ src/main/resources/templates/fragments/navbar.html
@@ -28,15 +28,15 @@
              </span>
                     <span sec:authorize="!isAuthenticated()" class="nav navbar-nav navbar-right">
                         <li><a href="#login-modal" data-toggle="modal" data-target="#login-modal">Log-in</a></li>
-                        <li><a th:href="@{/register}">Register</a></li>
-                        <!--<li><a href="#register-modal" data-toggle="modal" data-target="#register-modal">Register</a></li>-->
+                        <!--<li><a th:href="@{/register}">Register</a></li>-->
+                        <li><a href="#register-modal" data-toggle="modal" data-target="#register-modal">Register</a></li>
              </span>
                 </ul>
             </div>
         </nav>
 
         <div th:replace="fragments/login-modal :: login-modal"></div>
-        <!--<div th:replace="fragments/register-modal :: register-modal"></div>-->
+        <div th:replace="fragments/register-modal :: register-modal"></div>
 
     </div><!-- end of fragment div -->
